+++ src/features/club/api/schemas.ts
@@ -1,8 +1,23 @@
-export type ClubListResponse = {
-  clubId: number;
+/** ### 모임 생성
+ * - clubName: 모임명 (string)
+ * - description: 모임 소개글 (string)
+ * - category: 관심사 카테고리 (string)
+ * - region: 지역명 (string)
+ * - imgId?: 대표 이미지 ID (선택)
+ */
+export interface CreateClubRequest {
   clubName: string;
+  description: string;
   category: string;
-}[];
+  region: string;
+  imgId?: number;
+}
+
+/** ### 생성된 모임 ID
+ */
+export interface CreateClubResponse {
+  clubId: number;
+}
 
 export type ClubDetailResponse = {
   clubName: string;
@@ -11,3 +26,9 @@ export type ClubDetailResponse = {
   region: string;
   imageUrl: string;
 };
+
+export type ClubListResponse = {
+  clubId: number;
+  clubName: string;
+  category: string;
+}[];
